Use reflection to get path relative to implementing plugin
Passing a relative path to ::addUI would attempt to load a file relative to the directory of AllSpark.class.php and not the implementing child class. Not a problem if there's only one AllSpark-deriving plugin available, but a potential issue if there's more than one active in an install. Now uses reflection to get the file path to the implementing child.